# clients/openapi.yaml
openapi: 3.1.0
info:
  title: mem-dog API
  version: 4.0.0
  description: >
    Private AI system for multi-channel data ingestion, AI enrichment,
    temporal knowledge graph, and multi-signal semantic search.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key (md_*), JWT, or global API key
  schemas:
...
    EntityType:
      type: string
      enum: [person, organization, product, location, date, url, concept, event]
...
    Entity:
      type: object
      properties:
        entity_id:
          type: string
        name:
          type: string
        entity_type:
          $ref: '#/components/schemas/EntityType'
        canonical_name:
          type: string
        user_id:
          type: string
...
paths:
...
  /api/v1/channel-identities:
    post:
      tags: [Channels]
      summary: Create or upsert channel identity
      operationId: createChannelIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
  /api/v1/channel-identities/by-channel:
    get:
      tags: [Channels]
      summary: Get identity by channel
      operationId: getChannelIdentity
      parameters:
        - name: channel_type
          in: query
          required: true
          schema:
            type: string
        - name: channel_unique_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Identity
    patch:
      tags: [Channels]
      summary: Update channel identity
      operationId: updateChannelIdentity
      parameters:
        - name: channel_type
          in: query
          required: true
          schema:
            type: string
        - name: channel_unique_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [Channels]
      summary: Delete channel identity
      operationId: deleteChannelIdentity
      parameters:
        - name: channel_type
          in: query
          required: true
          schema:
            type: string
        - name: channel_unique_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /api/v1/channel-identities/by-user/{user_id}:
    get:
      tags: [Channels]
      summary: List identities for user
      operationId: listUserChannelIdentities
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Identities

  # ---------- Channels ----------
...
  /api/v1/channels/{channel_type}:
    get:
      tags: [Channels]
      summary: Get channel metadata
      operationId: getChannel
      parameters:
        - name: channel_type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel
    put:
      tags: [Channels]
      summary: Create or update channel
      operationId: updateChannel
      parameters:
        - name: channel_type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [Channels]
      summary: Delete channel
      operationId: deleteChannel
      parameters:
        - name: channel_type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ---------- AI Config ----------
...
  /api/v1/ai/analysis-templates:
    get:
      tags: [AnalysisTemplates]
      summary: List analysis templates
      operationId: listAnalysisTemplates
      parameters:
        - name: data_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Templates
    post:
      tags: [AnalysisTemplates]
      summary: Create template
      operationId: createAnalysisTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
...
  /api/v1/ai/analysis-templates/{template_id}:
    get:
      tags: [AnalysisTemplates]
      summary: Get template
      operationId: getAnalysisTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template
    put:
      tags: [AnalysisTemplates]
      summary: Update template
      operationId: updateAnalysisTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [AnalysisTemplates]
      summary: Delete template
      operationId: deleteAnalysisTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ---------- Agent Configs ----------
...
  /api/v1/graph/entities:
    get:
      tags: [Graph]
      summary: Search entities
      operationId: searchEntities
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: entity_type
          in: query
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entity'
...
  /api/v1/graph/entities/{entity_id}:
    get:
      tags: [Graph]
      summary: Get entity
      operationId: getEntity
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
    delete:
      tags: [Graph]
      summary: Delete entity
      operationId: deleteEntity
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...
  /api/v1/graph/data/{data_id}/entities:
    get:
      tags: [Graph]
      summary: Get entities extracted from data item
      operationId: getDataEntities
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Entities
    delete:
      tags: [Graph]
      summary: Delete data's entities
      operationId: deleteDataEntities
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...